REST path standards for Stock/LinkController

+++ src/main/java/home/holymiko/InvestmentScraperApp/Server/API/Repository/StockGrahamHistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockGrahamHistoryRepository extends JpaRepository<StockGrahamHistory, Integer> {
+public interface StockGrahamHistoryRepository extends JpaRepository<StockGrahamHistory, Long> {
 }
